Oprava chyb pocas testovania
Treba
-do FragmentMouseList nastavit vyhladavanie.
-do mouseView pridat zobrazenie fotky mysi
-zistit preco sa vo MouseVIew nezobrazuje zoznam predchadzajucich chyteni
-Odstranit runBlocking a nahradit LiveData aby sa neblokovalo hlavne vlakno.
-Pokracovat v testovani od Recapture

+++ app/src/main/java/com/example/datatrap/picture/fragments/TakePhotoFragment.kt
@@ -89,8 +89,10 @@ class TakePhotoFragment : Fragment(), EasyPermissions.PermissionCallbacks {
                     pictureViewModel.deletePicture(oldPicture)
                 }
                 // vymazat povodnu fotku ako subor
-                val myFile: File = File(oldPicture?.path)
-                if (myFile.exists()) myFile.delete()
+                if (oldPicture?.path != null) {
+                    val myFile: File = File(oldPicture.path)
+                    if (myFile.exists()) myFile.delete()
+                }
                 // ulozit novu
                 pictureViewModel.insertPicture(newPicture)
                 // zobrazit v galerii
